Progpow header size (#176)
* Updated block header and nonce size

* masking fix

* removed progpow fork height from testnet defaults

* version bump

* typo fix

* Updated nonce for ProgPoW calculation to reflect 40 bits in hashBlock(), added new NONCE_MASK to handle post-vBlake nonces with appropriate shift to preserve all 40 bits during block header assembly

* Updated block header chopping for ProgPoW calculation to account for 5-byte nonce.

* Fixed nonce assembly into block header in assembleBlockHeader()

* Added byteArrayToLong conversion

* Updated TestNetProgPoW starting Bitcoin block.

* Network checks configurability

* Fixes for header changes

* Fixed Long nonce extraction from ProgPoW block header

* Updated synthetic extra nonce to allow 40-bit nonces.

* 8 -> 6 hex characters for SYNTHETIC_EXTRA_NONCE

* 13 extra Keccak rounds for ProgPoW loop

* Updated ProgPoW Lanes to 32

* PROGPOW_LANES=15

* DAG_LOADS=8

* HASH_BYTES = 128

* Reduced seed size to 55 bits, updated PROGPOW_CNT_DAG to 128, PROGPOW_CNT_MATH to 20

* Updated Keccak THETA function

* Modified Keccakf800 algorithm

* Cleanup

* Updated Bitcoin starting block for ProgPoWTestNet

* merge fix

+++ nodecore-cli/src/main/java/nodecore/cli/serialization/BlockInfo.java
@@ -115,7 +115,7 @@ public class BlockInfo {
     public int encoded_difficulty;
 
     @SerializedName("winning_nonce")
-    public int winningNonce;
+    public long winningNonce;
 
     @SerializedName("bitcoin_block_headers")
     public List<String> bitcoinBlockHeaders = new ArrayList<>();
